Fixed devise current_use issue

+++ app/helpers/store_helper.rb
@@ -1,7 +1,7 @@
 module StoreHelper
 	def follow_button store
 
-		if  current_user != store.user 
+		if  current_user && current_user != store.user 
 			if current_user.following?(store, current_user)
 				button_to 'Unfollow', store_unfollow_path(store), method: :delete, class: "btn btn-danger button"
 			else
